Make token subject available to the security entity.
The token subject - even when it's not used for authorization as
part of XrdSecEntity's credential name - is potentially useful as
it indicates a unique entity.  Record this as part of the extended
attribute API in the security credential.

(cherry picked from commit 3eacd634950cdd4e349393b1a5a7a8f8c78caa10)