Add support for connection tracking in the throttle plugin.
This adds a limit for "max active connections" where we define the
'active connection' to be a thread associated with a unique entity
with at least one file open.

Admins would like to track & limit the number of open connections
because each connection incurs the cost of one OS thread and users
can overwhelm the server by having a large number of connections.

I consider the "active connections" a good enough approximation of
"open connections" because the admin can set the idle connections to
be reaped and we really only care about the resource in the coarse
grain: threads become expensive when measured in the thousands, not
in the dozens.

I also consider this a "good enough" approximation because it keeps
all the cost of tracking in the throttling plugin and not the framework
itself.  When not in use, there's effectively zero cost.

(cherry picked from commit 442a02a019b847d0e1807851796b32155cfa7cd6)