Have the XrdHttp extraction logic match GSI.
The GSI security protocol defaults to the "trymap" logic which,
according to documentation, is "try to map the DN but if unsuccessful
use the hash of the client’s DN as the user identifier (username)".
With this change, the XrdHttpSecurity interface will follow the same
logic.

Since some sites may have special setups which rely on the old mechanism,
one can get the old behavior by setting the optional new `compatNameGeneration`
configuration in the `http.gridmap` setting.  For example:

```
http.gridmap compatNameGeneration /etc/xrootd/grid-mapfile
```

Would restore the old behavior.

With this change, a user can effectively match the `nomap`, `trymap`,
and `usemap` settings between the XRootD and HTTP protocols and the
two protocols have the same default.  Without the change, only the
`usemap` setting could be matched with HTTP.  Having the defaults the
same greatly decreases the "surprise" factor of using both protocols;
the prior HTTP default of guessing a name from the DN is undocumented.

(cherry picked from commit ec0754f28fc3cffa551f5cdc42ecdb959b3a0830)